requirements.txt
asyncmy==0.2.9
Cython==3.0.12
faiss-cpu==1.9.0
fastapi==0.115.5
ffmpeg-python==0.2.0
greenlet==3.0.3
h5py==3.12.1
httplib2==0.22.0
jsonschema==4.23.0
langchain==0.3.9
langchain-community==0.3.9
langchain-openai==0.2.10
lxml==5.3.0
numpy==2.0.0
openai==1.69.0
openpyxl==3.1.5
pandas==2.2.3
pillow==10.4.0
proto-plus==1.25.0
protobuf==5.29.3
PyPDF2==3.0.1
python-dateutil==2.9.0
python-debian==0.1.49
python-dotenv==1.1.0
python-multipart==0.0.19
pydub==0.25.1
setuptools==78.1.0
tensorflow==2.19.0
tiktoken==0.8.0
timm==1.0.11
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
uvicorn[standard]==0.34.0
wheel==0.45.1
PyAudio==0.2.14
transformers==4.52.4
datasets==3.6.0
soundfile==0.13.1
speechrecognition==3.14.3
sounddevice==0.5.2
tf-keras==2.19.0
librosa==0.11.0
supabase==2.15.3